backend/requirements.txt
flask
flask-cors
pymongo
azure-storage-blob
gunicorn
python-dotenv

